gla_pal

Generates palettes using the GLA colours

gla_pal(gla_theme = NULL, palette_type = "categorical",
  main_colours = "mayoral", n = 6, muted = FALSE, inc0 = FALSE)

Arguments

gla_theme
One of "light" or "dark". If NULL will try to pick up current theme, or default to light, Default: NULL

palette_type
One of "categorical", "quantitative", "highlight" or "diverging", Default: 'categorical'

main_colours
One or more of "blue", "pink", "green", "red", "yellow", "orange", "purple" or "mayoral" as a string or list, Default: 'mayoral'

n
Number of colours in the palette. If palette_type = "Diverging", this is the number of colours on each side of the diverging scale (so 2n (+1) colours in total will be returned). If palette_type = "Highlight" gla_pal will return main_colours + (n - length(main_colours)) context colours. Default: 6

muted
boolean, If TRUE a muted version of the gla colours will be used, Default: FALSE

inc0
boolean, If TRUE an additional colour representing zero will be added to a diverging palette. This results in a palette of length 2n + 1, Default: FALSE

Details

DETAILS

See also

Examples

# NOT RUN {
if(interactive()){
 #EXAMPLE1
 }
# }