theme_gla

Controls the non-data display to be consistent with our style guide

theme_gla(
  gla_theme = "default",
  x_axis_title = FALSE,
  y_axis_title = FALSE,
  xgridlines = FALSE,
  legend = TRUE,
  base_size = 14,
  free_y_facets = FALSE,
  y_label_length = 100
)

Arguments

gla_theme
Either 'default' or 'inverse', Default: 'default'

x_axis_title
boolean, If TRUE an x-axis label/title can be added, Default: FALSE

y_axis_title
boolean, If TRUE a y-axis label/title can be added, Default: FALSE

xgridlines
boolean, If TRUE, theme will include gridlines on the x-axis, Default: FALSE

legend
boolean, If TRUE, theme will include a legend at the top left of the plot, Default: TRUE

base_size
base font size, Default: 14

free_y_facets
Use if faceting plot with free y scales. Adjusts position of y axis labels, Default: FALSE

y_label_length
length of lines under y-axis labels. For most plots this won't need changing. It will need adjusting for faceted plots or if labels are particularly long. As a rough guide this should be set approximately equal to 96 * strwidth(y_labels, unit = "inches") with default base_size.

Examples

if (FALSE) {
if(interactive()){
 #EXAMPLE1
 }
}